move voxel classes from glut to glfw #2
