+++ examples_glut/CMakeLists.txt
@@ -7,8 +7,5 @@ add_subdirectory(00_viewer3d)
 add_subdirectory(glbuffer)
 add_subdirectory(shader)
 
-# geometry
-add_subdirectory(edit_cubes)
-
 # fem
 add_subdirectory(rigidbody3d_rotation)
