+++ src_pybind/core/main.cpp
@@ -18,7 +18,8 @@
 #include "delfem2/mat3.h"
 #include "delfem2/primitive.h"
 #include "delfem2/mshtopoio.h"
-#include "delfem2/voxel.h"
+#include "delfem2/gridvoxel.h"
+#include "delfem2/gridcube.h"
 #include "delfem2/bv.h"
 #include "delfem2/iss.h"
 #include "delfem2/slice.h"
