+++ test_cpp/main.cpp
@@ -13,7 +13,8 @@
 #include "delfem2/vec3.h"
 #include "delfem2/mat3.h"
 #include "delfem2/quat.h"
-#include "delfem2/voxel.h"
+#include "delfem2/gridvoxel.h"
+#include "delfem2/gridcube.h"
 #include "delfem2/mshtopo.h"
 #include "delfem2/mshmisc.h"
 #include "delfem2/mshio.h"
